Use scroll_method/button on all pointers
Scroll wheel still works and its not likely that any other pointing
device I own will have a button 278 that would cause issues, so lets just
make it easier on myself. This way it does not matter if we connect via bluetooth,
USB, or the RF-USB it will always work.